Section 1: Game Setup
Created an Unreal Engine Plugin called Inventory, complete with its own player controller and HUD widget. Created a sample item (potion) with a custom static mesh that allows the item to be highlighted on the custom Item Trace channel. The mesh also displays a pickup message when highlighted and removes it when it's not. Before any coding, I imported all the assets we would need for this project. Next commit will focus on the Inventory HUD.