Section 2: Inventory HUD
Created an Inventory Component and a Grid Frame that will be used to place items into inventory when acquired by the player. There are 3 separate grids: equippables, consumables, and craftables, each with their own button and display. Grids operate on a widget switcher (yup, that's a thing)  to display the separate grids when the  corresponding button is clicked. The next commit will focus on inventory data.